requirements.txt
editdistance==0.5.2
lmdb==1.0.0
matplotlib==3.2.1
numpy==1.19.5
opencv-python==4.4.0.46
path==15.0.0
tensorflow==2.4.0
Pillow~=8.2.0
pip~=21.1.3
wheel~=0.36.2
setuptools~=56.0.0